# docs/swagger.yaml
basePath: /
definitions:
...
  dto.Metrics:
    properties:
      delta:
        description: Delta значение метрики в случае передачи counter
        type: integer
      hash:
        description: Hash значение хеш-функции - подпись для проверки подлинности
          метрики
        type: string
      id:
        description: ID имя метрики
        type: string
      type:
        description: MType параметр, принимающий значение gauge или counter
        type: string
      value:
        description: Value значение метрики в случае передачи gauge
        type: number
    type: object
host: localhost:8080
info:
  contact: {}
  description: Сервис хранения метрик и счетчиков.
  title: Metrics storage API
  version: "1.0"
paths:
...
  /value/{metric_type}/{metric_name}:
    get:
      parameters:
      - description: 'Тип метрики: counter, gauge'
        in: path
        name: metric_type
        required: true
        type: string
      - description: Имя метрики
        in: path
        name: metric_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Metrics'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Запрос одной метрики с указанием её типа и имени
      tags:
      - Metrics
...
swagger: "2.0"
...